Refactor preprocessing and analysis modules
